requirements.txt
pytest==8.1.1
selenium==4.19.0
webdriver_manager==4.0.1
allure-pytest==2.13.4
pytest-rerunfailures==14.0
pytest-xdist==3.5.0
Autologging==1.3.2
assertpy==1.1
Faker==24.4.0
tenacity==8.2.3
pylint==3.1.0

